# Optimize Decimal.Remainder