In addition to better performance this contributes to making decimal a readonly struct.
I'll add new tests to CoreFX also.
Contributes to https://github.com/dotnet/coreclr/issues/18249.
### x64
|  Method |      Mean |     Error |    StdDev | Scaled |
|-------- |----------:|----------:|----------:|-------:|
|  Native | 169.42 ns | 0.3258 ns | 0.1162 ns |   1.00 |
| CoreRT2 |  78.66 ns | 0.0869 ns | 0.0310 ns |   0.46 |
### x86
|  Method |     Mean |     Error |    StdDev | Scaled |
|-------- |---------:|----------:|----------:|-------:|
|  Native | 128.0 ns | 0.1594 ns | 0.0568 ns |   1.00 |
| CoreRT2 | 120.4 ns | 0.2057 ns | 0.0734 ns |   0.94 |

@jkotas I have a WIP [readonly decimal branch](https://github.com/pentp/corert/commit/ee175a15451c95eddb5941b720219b0b13d47ba5), it doesn't compile yet because of `VarDecMod` and `NumberBufferToDecimal`, but can you check if I'm on the right track?