@pentp 2018-06-11
I added new tests that cover all `Decimal.Remainder` code paths now, but I might have been a bit too eager with fixing the `NearNegativeZero/NearPositiveZero` scale issue (https://github.com/dotnet/coreclr/issues/12605) because now CoreRT and CoreFX return different values, but the tests are shared.
Should I add back the scale bug for CoreRT or fix it in CoreCLR?