@jkotas 2018-06-11
Is fixing it just about changing the scale from 27 to 28? I think it would be easier to fix it.