@pentp 2018-06-11
The fix is in https://github.com/dotnet/coreclr/pull/18424 and I updated tests also (tests will fail until until CoreCLR update, NETFX is excluded).